Add new LongConstantString Codec behind feature flag
The new codec is enabled for reading but odc will not write data with the new codec unless "ODC_ENABLE_WRITING_LONG_STRING_CODEC" is set in the environment. This is to give some time for installations in the wild to update before files containing the new code are produced by default.